MiniWall Software
The MiniWall software is browser based software that is used to analyze
and understand large and small computationally and experimentally
generated data sets. The MiniWall software presents a matrix view of
plots and other visualization images and provides a responsive and
productive way to investigate the data and extract knowledge and
understanding from it. Since it is browser based, it runs on any
computer or device that can display a web page.